EvObj: select object with double touch now
EvShell: BitmapTransform interpretation bug

Global: standardization SetValue input parameter to int32_t

EvEve: BitmapLayout input parameters changed, BitmapTransform Coeff changed to s15f16 format and, optimized Vertex2f

EvImage: first release, RotateAround support on all EVE versions but ScaleX and ScaleY should be equal at this time

+++ src/code/EvButton.cpp
@@ -142,7 +142,7 @@ void        EvButton::SetOnTouch(void (*OnTouch)(EvButton *Sender, EvTouchEvent 
  * 
  * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
 
-void        EvButton::SetOnChange(void (*OnChange)(EvButton *Sender, int16_t Value))
+void        EvButton::SetOnChange(void (*OnChange)(EvButton *Sender, int32_t Value))
 {
   mOnChange = OnChange;
 }
